Move `cmd/influx_stress/runner` pkg to `stress`
