+++ stress/runner.go
